ai-core-witraz/requirements.txt
google-genai
google-cloud-storage
google-cloud-firestore
functions-framework